requirements.txt
altgraph==0.17.4
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
macholib==1.16.3
MarkupSafe==3.0.2
mecab-python3==1.0.9
packaging==24.1
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
setuptools==75.3.0
Werkzeug==3.0.6
